# contracts/openapi/plugin5.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "apiKey": {
        "description": "Development and service API key. Not for production traffic.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "oauth2_authorization_code_pkce": {
        "description": "Interactive user login with PKCE.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://id.pas.example/authorize",
            "refreshUrl": "https://id.pas.example/token",
            "scopes": {
              "admin:read": "Grants admin:read.",
              "admin:write": "Grants admin:write.",
              "data:read": "Grants data:read.",
              "data:write": "Grants data:write.",
              "distribution:kyc": "Grants distribution:kyc.",
              "distribution:publish": "Grants distribution:publish.",
              "distribution:read": "Grants distribution:read.",
              "distribution:write": "Grants distribution:write.",
              "ifrs17:read": "Grants ifrs17:read.",
              "ifrs17:write": "Grants ifrs17:write.",
              "insurance:read": "Grants insurance:read.",
              "insurance:write": "Grants insurance:write.",
              "ledger:read": "Grants ledger:read.",
              "ledger:write": "Grants ledger:write.",
              "products:publish": "Grants products:publish.",
              "products:read": "Grants products:read.",
              "products:write": "Grants products:write.",
              "uw:kyc": "Grants uw:kyc.",
              "uw:override": "Grants uw:override.",
              "uw:read": "Grants uw:read.",
              "uw:submit": "Grants uw:submit.",
              "workflow:write": "Grants workflow:write."
            },
            "tokenUrl": "https://id.pas.example/token"
          }
        },
        "type": "oauth2"
...
  },
  "info": {
    "description": "Embedded Insurance Distribution.\n\n**Problem.** Carriers have spent two decades building point-of-sale platforms partners cannot use, and most distribution still runs on a phone call and a PDF. A partner integrating a carrier's platform should not need a bespoke integration per carrier.\n\n**Approach.** Distribution as an API. A partner gets a licensed catalog filtered by their own entitlement, quotes that name the premium *and* what they earn, onboarding with suitability and KYC gating, payment intents, and a commission statement that claws back inside free look. The sequence is the same one a carrier already runs - lead, quote, suitability, KYC, bind, pay, commission - so a partner is not learning a second business model.\n\n**The money rule.** Nothing here is a ledger of record. An intent records that money was requested; a payment records that it arrived; the carrier's finance system stays authoritative. `reconcile` exists so both sides compare against the same figures rather than each guessing.\n\nREST endpoints live under `/distribution`. the same fourteen capabilities are exposed as MCP tools over streamable HTTP at `/mcp`.",
    "title": "plugin5 - PAS Plugins",
    "version": "1.0.0",
    "x-acord-ngds": {
      "sourceRef": "https://www.acord.org/standards-architecture/acord-data-standards/next-generation-digital-standards",
      "standardVersion": "6.0.0",
      "transactionCodes": 40
    },
    "x-mcp": {
      "path": "/mcp",
...
  "openapi": "3.1.0",
  "paths": {
...
    "/distribution/catalog": {
      "get": {
        "description": "Products licensed to a partner, filtered by entitlement and jurisdiction.",
        "operationId": "distribution.catalog",
        "parameters": [
          {
            "in": "query",
            "name": "partnerId",
            "required": true,
            "schema": {
              "minLength": 2,
              "title": "Partnerid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Distribution.Catalog",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Products this partner may sell",
        "tags": [
          "embedded-distribution"
        ]
      }
    },
...
    "/distribution/intents/{intent_id}/capture": {
      "post": {
        "description": "Capture a payment intent. This records an arrival; it is not a ledger.",
        "operationId": "distribution.capture",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Distribution.Capture",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record that money arrived",
        "tags": [
          "embedded-distribution"
        ]
      }
    },
...
  },
  "security": [
    {
      "oauth2_authorization_code_pkce": [],
      "tenantHeader": []
    },
    {
      "oauth2_client_credentials": [],
      "tenantHeader": []
    },
    {
      "apiKey": [],
      "tenantHeader": []
    }
  ],
...
}